# backend/openapi.json
{
  "components": {
    "schemas": {
      "AgreementResult": {
        "properties": {
          "consensus_midpoint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Midpoint"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgreementRow"
            },
            "title": "Items",
            "type": "array"
          },
          "reference_longitude": {
            "title": "Reference Longitude",
            "type": "number"
          },
          "resolution_count": {
            "title": "Resolution Count",
            "type": "integer"
          }
        },
        "required": [
          "country",
          "resolution_count",
          "reference_longitude",
          "consensus_midpoint",
          "items"
        ],
        "title": "AgreementResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MultilateralResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MultilateralRow"
            },
            "title": "Items",
            "type": "array"
          },
          "mean_alignment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Alignment"
          },
          "minimum_votes": {
            "title": "Minimum Votes",
            "type": "integer"
          },
          "resolution_count": {
            "title": "Resolution Count",
            "type": "integer"
          }
        },
        "required": [
          "resolution_count",
          "minimum_votes",
          "mean_alignment",
          "items"
        ],
        "title": "MultilateralResult",
        "type": "object"
      },
...
      "SubjectResult": {
        "properties": {
          "comparison": {
            "title": "Comparison",
            "type": "string"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SubjectScore"
            },
            "title": "Items",
            "type": "array"
          },
          "minimum_shared_votes": {
            "title": "Minimum Shared Votes",
            "type": "integer"
          }
        },
        "required": [
          "country",
          "comparison",
          "minimum_shared_votes",
          "items"
        ],
        "title": "SubjectResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only access to adopted UN General Assembly resolutions and the existing Policy Pulse methodology. Scores are 0\u20131; unavailable observations are null.",
    "title": "Policy Pulse API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis/agreement/{country}": {
      "get": {
        "operationId": "agreement_api_v1_analysis_agreement__country__get",
        "parameters": [
          {
            "in": "path",
            "name": "country",
            "required": true,
            "schema": {
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "title": "Subject",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "include_descendants",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Descendants",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "country_mode",
            "required": false,
            "schema": {
              "default": "none",
              "enum": [
                "none",
                "voted",
                "member"
              ],
              "title": "Country Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 300,
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgreementResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agreement",
        "tags": [
          "Analysis"
        ]
      }
    },
    "/api/v1/analysis/multilateral": {
      "get": {
        "operationId": "multilateral_api_v1_analysis_multilateral_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "title": "Subject",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "include_descendants",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Descendants",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "country_mode",
            "required": false,
            "schema": {
              "default": "none",
              "enum": [
                "none",
                "voted",
                "member"
              ],
              "title": "Country Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 300,
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultilateralResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Multilateral",
        "tags": [
          "Analysis"
        ]
      }
    },
    "/api/v1/analysis/subjects/{country}/{comparison}": {
      "get": {
        "operationId": "subjects_api_v1_analysis_subjects__country___comparison__get",
        "parameters": [
          {
            "in": "path",
            "name": "country",
            "required": true,
            "schema": {
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "comparison",
            "required": true,
            "schema": {
              "title": "Comparison",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Subjects",
        "tags": [
          "Analysis"
        ]
      }
    },
    "/api/v1/analysis/timeline/{country}": {
      "get": {
        "description": "Full-history session means. Date/subject filters intentionally do not apply.",
        "operationId": "timeline_api_v1_analysis_timeline__country__get",
        "parameters": [
          {
            "in": "path",
            "name": "country",
            "required": true,
            "schema": {
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "compare",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 50,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare"
            }
          },
          {
            "in": "query",
            "name": "include_special",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Special",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline",
        "tags": [
          "Analysis"
        ]
      }
    },
    "/api/v1/analysis/words/{mode}": {
      "get": {
        "operationId": "words_api_v1_analysis_words__mode__get",
        "parameters": [
          {
            "in": "path",
            "name": "mode",
            "required": true,
            "schema": {
              "enum": [
                "default",
                "geopolitical",
                "thematic",
                "action",
                "category"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "title": "Subject",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "include_descendants",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Descendants",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "country_mode",
            "required": false,
            "schema": {
              "default": "none",
              "enum": [
                "none",
                "voted",
                "member"
              ],
              "title": "Country Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 300,
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Words",
        "tags": [
          "Analysis"
        ]
      }
    },
    "/api/v1/countries/{country}/profile": {
      "get": {
        "operationId": "profile_api_v1_countries__country__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "country",
            "required": true,
            "schema": {
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "compare",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 50,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Profile",
        "tags": [
          "Countries"
        ]
      }
    },
    "/api/v1/metadata": {
      "get": {
        "operationId": "metadata_api_v1_metadata_get",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "en",
              "enum": [
                "en",
                "fr",
                "es",
                "ar",
                "zh",
                "ru"
              ],
              "title": "Language",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Metadata",
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/resolutions": {
      "get": {
        "operationId": "resolutions_api_v1_resolutions_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "description": "Date between 1678-01-01 and 2261-12-31, inclusive.",
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "title": "Subject",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "include_descendants",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Descendants",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "country_mode",
            "required": false,
            "schema": {
              "default": "none",
              "enum": [
                "none",
                "voted",
                "member"
              ],
              "title": "Country Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 300,
              "title": "Keyword",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "compare",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 50,
              "title": "Compare",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "vote",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Y",
                    "N",
                    "A",
                    "X"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vote"
            }
          },
          {
            "in": "query",
            "name": "agreement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "AGREED",
                    "DISAGREED",
                    "STRONGLY_DISAGREED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agreement"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "date_desc",
              "enum": [
                "date_desc",
                "date_asc",
                "consensus_desc",
                "consensus_asc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolutionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolutions",
        "tags": [
          "Resolutions"
        ]
      }
    },
...
    "/api/v1/resolutions/{resolution_id}": {
      "get": {
        "operationId": "resolution_detail_api_v1_resolutions__resolution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resolution_id",
            "required": true,
            "schema": {
              "title": "Resolution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolutionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolution Detail",
        "tags": [
          "Resolutions"
        ]
      }
    },
...
  }
}